Initial Commit + AI and Cameras
Uploading project:

AI has the following:

Go to reported location when a camera detects the player (which is then reported 1.5 seconds later if the player is still in sight)

Investigate the surrounding area then return to the reported location with a random amount of attempts

Return to AI spawn location

Camera features include:

Delayed Report Feature: The camera will only report the player if they remain in sight for 1.5 seconds

Instance Editable Variables that allow for control on which guards it should affect, or if all guards should be affected.

Tracks player real-time location but with limited accuracy based on time delay.